requirements.txt
redis>=4.0.0
requests>=2.28.0
fastmcp>=0.1.0
llama-cpp-python>=0.3.0
huggingface_hub>=0.20.0
openai-whisper>=20231117
uvicorn>=0.20.0
starlette>=0.27.0
stable-diffusion-cpp-python>=0.1.0
duckduckgo-search>=8.0.0
trafilatura>=1.6.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
textblob>=0.18.0
rapidfuzz>=3.0.0
markdown>=3.5.0
langdetect>=1.0.9
sumy>=0.11.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
pyyaml>=6.0.0
tabulate>=0.9.0
sympy>=1.12
pint>=0.23
pillow>=10.0.0
qrcode>=7.4
python-barcode>=0.15.0
psutil>=5.9.0
python-jose>=3.3.0
apprise>=1.7.0
icalendar>=5.0.0
python-dateutil>=2.8.0
croniter>=1.4.0
tinydb>=4.8.0
python-slugify>=8.0.0
jinja2>=3.1.0
gitpython>=3.1.0
chardet>=5.2.0
colorthief>=0.2.1
anthropic>=0.40.0
openai>=1.0.0
alembic>=1.12.0
sqlalchemy>=2.0.0
websockets>=12.0
